feat(action): Use dynamic image tag for GitHub Action
Update `action.yml` to use `github.action_ref` as the Docker image tag. This ensures that the GitHub Action runs with a version of the custom MegaLinter flavor image that corresponds to the Git reference (branch or tag) where the action is being used. This improves versioning consistency and reproducibility.

Additionally, this commit includes minor style and documentation improvements:
- Removed unnecessary quotes from string values.
- Changed Docker argument `-v` to `--volume` for clarity.
- Added a reference comment to the original MegaLinter template.